requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
lxml==4.5.1
numpy==1.16.6
opencv-python==4.2.0.32
Pillow==6.2.2
pytesseract==0.3.4
requests==2.23.0
tesseract==0.1.3
urllib3==1.25.9
